Refactoring calendar, docs + registry
