+++ resources/js/docs/components/date-and-time/calendar.mdx
@@ -1,0 +1,32 @@
+---
+title: "Calendar"
+description: "A calendar displays one or more date grids and allows users to select a single date."
+---
+
+
+
+
+<DemoCodePreview codePath="calendar/default.tsx" demoPath="demos/calendar/default" className="h-120" />
+
+
+## Installation
+
+<InstallationSection
+  dependencies={['react-aria-components', '@internationalized/date', 'react-aria', 'tailwind-variants']}
+  components={['button.tsx', 'select.tsx']}
+>
+  <InstallationCommand command="execute" params="shadcn@latest add @twc-ui/calendar" />
+</InstallationSection>
+
+
+## Documentation
+
+### Props
+
+[React Aria Calendar props](https://react-spectrum.adobe.com/react-aria/Calendar.html#calendar-1)
+
+| Prop          | Type                                                                          | Default | Description                                                          |
+|---------------|-------------------------------------------------------------------------------|---------|----------------------------------------------------------------------|
+| footerButtons | <TypeInfo title="FooterButtons" types={['both', 'none', 'reset', 'today']} /> | `both`  | Display footer buttons                                               |
+| maxYears      | `number`                                                                      | `100`   | If minValue and maxValue are not set, max years for the years select |
+| errorMessage  | `string`                                                                      | ``      | Error message under the calendar                                     |
