+++ resources/js/docs/getting-started/introduction.mdx
@@ -1,6 +1,6 @@
 ---
 title: Introduction
-description: DRY, more accessible shadcn/ui-alike React Aria components, optimized for Inertia.js with Laravel's Precognition feature build in.
+description: DRY-er, more accessible shadcn/ui-alike React Aria components, optimized for Inertia.js with Laravel's Precognition feature build in.
 order: 10
 ---
 ## Motivation
@@ -16,6 +16,12 @@ feature.
 
 ### JollyUI
 Specifically, the form components are influenced by [jolbol1's](https://github.com/jolbol1) **[jolly-ui](https://www.jollyui.dev/)**. Although my library only contains a few React Aria components, his library supports the entire Adobe library.
+
+### Taki UI
+[Taki UI](https://taki-ui.com/) is a fork of shadcn/ui rebuilt with React Aria Components instead of Radix UI.
+
+### Intent UI
+[Intent UI](https://intentui.com/) is a chill set of React components, built on top of React Aria Components
 
 ### PaceUI
 Initially, the library and documentation used [Pace UI](https://www.paceui.com/) as scaffolding. While this project is now realized with the React Starter kit of Laravel, some of his documentation components are furthermore the base of mine.
