+++ resources/js/pages/demos/calendar/default.tsx
@@ -1,0 +1,10 @@
+import { DemoContainer } from '@/components/docs/DemoContainer'
+import { Calendar } from '@/components/twc-ui/calendar'
+
+export default function Dashboard() {
+  return (
+    <DemoContainer>
+      <Calendar />
+    </DemoContainer>
+  )
+}
